chore: make requested changes
