+++ packages/api-extractor-utils/src/index.ts
@@ -1,6 +1,4 @@
 export * from './ApiNodeJSONEncoder';
-export * from './indexGenerator';
 export * from './parse';
-export * from './indexGenerator';
 export * from './tsdoc';
 export * from './TypeParameterJSONEncoder';
