+++ packages/scripts/.gitignore
@@ -23,3 +23,5 @@ docs/**/*
 # Miscellaneous
 .tmp/
 coverage/
+
+searchIndex/
