+++ packages/website/tsconfig.json
@@ -5,14 +5,14 @@
 		"isolatedModules": true,
 		"jsx": "preserve",
 		"baseUrl": ".",
-		"allowJs": false,
 		"noEmit": true,
+		"allowJs": false,
 		"incremental": true,
 		"skipLibCheck": true,
 		"paths": {
 			"~/*": ["./src/*"]
 		}
 	},
-	"include": ["src/**/*.ts", "src/**/*.tsx", "src/**/*.mts", "next-env.d.ts", "types.d.ts", "src/index.ts"],
+	"include": ["src/**/*.ts", "src/**/*.tsx", "next-env.d.ts", "types.d.ts"],
 	"exclude": ["node_modules"]
 }
